mismatched types resolved

+++ programs/nonce_savings/src/state.rs
@@ -9,7 +9,7 @@ pub struct SavingsAccount {
     pub sol_balance: u64,             // SOL balance in the savings account
     pub usdc_balance: u64,            // USDC balance in the savings account
     pub current_time: i64,            // Timestamp when the savings account was created
-    pub lock_duration: u16,           // Lock duration (in days)
+    pub lock_duration: i64,           // Lock duration (in days)
     pub usd_price: Option<f64>,       // Optional field to store USD price
     pub bump: u8,                     // Bump seed for PDA (program-derived account)
     pub type_of_savings: SavingsType, // Type of savings (enum)
